GameSpace: Don't add floating action if game closed too quickly
We use handler.postDelayed() to draw the floating action with
500ms delay on starting game.
If user opens and closes the game too quickly,
the game is getting closed and called onGameLeave()
before the handler.postDelayed execution so the floating
action is drawn even after closing the game.

Even the worse, user can't close that floating action without
reboot and can create infinite floating actions.

So don't draw the floating action if game closed before that.

Change-Id: I7fc3748137ff6d3fdb71b8b3590580963f4d0c96